# libs/api-client/generated-api-docs.json
{"openapi":"3.0.0","paths":{"/user/{id}":{"get":{"operationId":"user","summary":"","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"401":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnauthorizedException"}}}}},"tags":["user"]},"post":{"operationId":"updateUser","summary":"","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserRequest"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"401":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnauthorizedException"}}}}},"tags":["user"]},"delete":{"operationId":"removeUser","summary":"","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""},"401":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnauthorizedException"}}}}},"tags":["user"]}},"/user":{"get":{"operationId":"users","summary":"","parameters":[{"name":"order","required":false,"in":"query","schema":{"default":"ASC","enum":["ASC","DESC"],"type":"string"}},{"name":"page","required":false,"in":"query","schema":{"minimum":1,"default":1,"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"minimum":1,"maximum":50,"default":10,"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/PageDto"},{"properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/UserDto"}}}}]}}}},"401":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnauthorizedException"}}}}},"tags":["user"]},"post":{"operationId":"createUser","summary":"","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserRequest"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"401":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnauthorizedException"}}}}},"tags":["user"]}},"/todo/{id}":{"get":{"operationId":"todo","summary":"","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TodoDto"}}}},"401":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnauthorizedException"}}}}},"tags":["todo"]},"post":{"operationId":"updateTodo","summary":"","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTodoRequest"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TodoDto"}}}},"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TodoDto"}}}},"401":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnauthorizedException"}}}}},"tags":["todo"]},"delete":{"operationId":"removeTodo","summary":"","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""},"401":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnauthorizedException"}}}}},"tags":["todo"]}},"/todo":{"get":{"operationId":"todos","summary":"","parameters":[{"name":"order","required":false,"in":"query","schema":{"default":"ASC","enum":["ASC","DESC"],"type":"string"}},{"name":"page","required":false,"in":"query","schema":{"minimum":1,"default":1,"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"minimum":1,"maximum":50,"default":10,"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/PageDto"},{"properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/TodoDto"}}}}]}}}},"401":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnauthorizedException"}}}}},"tags":["todo"]},"post":{"operationId":"createTodo","summary":"","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTodoRequest"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TodoDto"}}}},"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TodoDto"}}}},"401":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnauthorizedException"}}}}},"tags":["todo"]}},"/auth/login":{"post":{"operationId":"login","summary":"","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"401":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnauthorizedException"}}}}},"tags":["auth"]}},"/auth/logout":{"post":{"operationId":"logout","summary":"","parameters":[],"responses":{"200":{"description":""},"201":{"description":""},"401":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnauthorizedException"}}}}},"tags":["auth"]}},"/auth/me":{"get":{"operationId":"me","summary":"","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}},"tags":["auth"]}},"/health":{"get":{"operationId":"check","parameters":[],"responses":{"200":{"description":"The Health Check is successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"ok"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"error":{"type":"object","example":{},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}}}}}}}},"503":{"description":"The Health Check is not successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"error"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"error":{"type":"object","example":{"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"},"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}}}}}}}}},"tags":["health"]}}},"info":{"title":"api","description":"API docs","version":"1.0","contact":{}},"tags":[{"name":"public","description":""}],"servers":[],"components":{"schemas":{"UserDto":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"role":{"enum":["user","admin"],"type":"string"}},"required":["id","name","role"]},"UnauthorizedException":{"type":"object","properties":{}},"PageMetaDto":{"type":"object","properties":{"page":{"type":"number"},"limit":{"type":"number"},"itemCount":{"type":"number"},"pageCount":{"type":"number"},"hasPreviousPage":{"type":"boolean"},"hasNextPage":{"type":"boolean"}},"required":["page","limit","itemCount","pageCount","hasPreviousPage","hasNextPage"]},"PageDto":{"type":"object","properties":{"meta":{"$ref":"#/components/schemas/PageMetaDto"},"data":{"type":"array","items":{"type":"string"}}},"required":["meta","data"]},"CreateUserRequest":{"type":"object","properties":{"name":{"type":"string"},"role":{"type":"string","enum":["user","admin"]}},"required":["name","role"]},"UpdateUserRequest":{"type":"object","properties":{"name":{"type":"string"}},"required":["name"]},"TodoDto":{"type":"object","properties":{"id":{"type":"string"},"title":{"type":"string"},"user":{"$ref":"#/components/schemas/UserDto"}},"required":["id","title","user"]},"CreateTodoRequest":{"type":"object","properties":{"title":{"type":"string"}},"required":["title"]},"UpdateTodoRequest":{"type":"object","properties":{"title":{"type":"string"}},"required":["title"]},"LoginDto":{"type":"object","properties":{"name":{"type":"string"}},"required":["name"]}}}}